refactor: move slides into page slice
