+++ src/pages/slides/entities/Slide/lib/slideBuilder.ts
@@ -1,5 +1,5 @@
-import type { CodeLine } from '../../CodeLine/model/types';
-import type { SourceFileNode } from '../../SourceFileNode/model/types';
+import type { CodeLine } from '@/entities/CodeLine/model/types';
+import type { SourceFileNode } from '@/entities/SourceFileNode/model/types';
 import type { Slide, SlideContext } from '../model/types';
 
 /**
