+++ src/pages/slides/features/SlideNavigation/model/atoms.ts
