+++ src/pages/slides/widgets/SlideView/SlideView.tsx
@@ -1,8 +1,8 @@
 import { useAtomValue, useSetAtom } from 'jotai';
 import { useEffect } from 'react';
-import { graphDataAtom } from '../../entities/AppView/model/atoms';
+import { graphDataAtom } from '@/entities/AppView/model/atoms';
+import { activeTabAtom, openedTabsAtom } from '@/features/File/OpenFiles/model/atoms';
 import { buildSlidesFromFiles } from '../../entities/Slide/lib/slideBuilder';
-import { activeTabAtom, openedTabsAtom } from '../../features/File/OpenFiles/model/atoms';
 import { currentSlideAtom, currentSlideIdAtom, slidesAtom } from '../../features/SlideNavigation/model/atoms';
 import SlideContent from './SlideContent';
 import SlideContext from './SlideContext';
